+++ redactions/user.js
@@ -11,11 +11,6 @@ import {
   deleteArea
 } from 'services/areas';
 import {
-  deleteFavourite,
-  createFavourite,
-  fetchFavourites
-} from 'services/favourites';
-import {
   fetchSubscriptions,
   deleteSubscription
 } from 'services/subscriptions';
@@ -26,11 +21,6 @@ import { fetchGeostore, fetchCountry } from 'services/geostore';
  * CONSTANTS
 */
 const SET_USER = 'user/setUser';
-// favourites
-const SET_USER_FAVOURITES = 'user/setUserFavourites';
-const SET_USER_FAVOURITES_LOADING = 'user/setUserFavouritesLoading';
-const SET_USER_FAVOURITES_FIRST_LOAD = 'user/setUserFavouritesFirstLoad';
-const SET_USER_FAVOURITES_ERROR = 'user/setUserFavouritesError';
 // areas
 const SET_USER_AREAS = 'user/setUserAreas';
 const SET_USER_AREAS_ERROR = 'user/setUserAreasError';
@@ -41,12 +31,6 @@ const SET_USER_AREA_LAYER_GROUP = 'user/setUserAreaLayerGroup';
  * REDUCER
 */
 const initialState = {
-  favourites: {
-    items: [],
-    loading: false,
-    error: null,
-    isFirstLoad: false
-  },
   areas: {
     items: [],
     layerGroups: {},
@@ -59,46 +43,6 @@ export default function (state = initialState, action) {
   switch (action.type) {
     case SET_USER: {
       return { ...state, ...action.payload };
-    }
-
-    case SET_USER_FAVOURITES: {
-      return {
-        ...state,
-        favourites: {
-          ...state.favourites,
-          items: action.payload
-        }
-      };
-    }
-
-    case SET_USER_FAVOURITES_LOADING: {
-      return {
-        ...state,
-        favourites: {
-          ...state.favourites,
-          loading: action.payload
-        }
-      };
-    }
-
-    case SET_USER_FAVOURITES_FIRST_LOAD: {
-      return {
-        ...state,
-        favourites: {
-          ...state.favourites,
-          isFirstLoad: action.payload
-        }
-      };
-    }
-
-    case SET_USER_FAVOURITES_ERROR: {
-      return {
-        ...state,
-        favourites: {
-          ...state.favourites,
-          error: action.payload
-        }
-      };
     }
 
     case SET_USER_AREAS: {
@@ -167,66 +111,6 @@ export function setUser(user) {
     dispatch({ type: SET_USER, payload: userObj });
   };
 }
-
-// FAVOURITES
-export const setFavouriteLoading = createAction(SET_USER_FAVOURITES_LOADING);
-export const setFavouriteFirstLoad = createAction(SET_USER_FAVOURITES_FIRST_LOAD);
-export const setFavouriteError = createAction(SET_USER_FAVOURITES_ERROR);
-
-export const getUserFavourites = createThunkAction('user/getUserFavourites', () =>
-  (dispatch, getState) => {
-    const { token } = getState().user;
-
-    if (!token) {
-      return;
-    }
-
-    dispatch(setFavouriteLoading(true));
-    fetchFavourites(token)
-      .then((data) => {
-        dispatch(setFavouriteLoading(false));
-        dispatch(setFavouriteFirstLoad(true));
-        dispatch({ type: SET_USER_FAVOURITES, payload: data });
-      })
-      .catch((error) => {
-        dispatch(setFavouriteLoading(false));
-        dispatch(setFavouriteError(error));
-        dispatch({ type: SET_USER_FAVOURITES, payload: [] });
-      });
-  });
-
-export const toggleFavourite = createThunkAction('user/toggleFavourite', (payload = {}) =>
-  (dispatch, getState) => {
-    const { token } = getState().user;
-    const { favourite, resource } = payload;
-
-    dispatch(setFavouriteLoading(true));
-
-    if (favourite.id) {
-      const { id } = favourite;
-      deleteFavourite(token, id)
-        .then(() => {
-          // asks for the new updated list of favourites
-          dispatch(getUserFavourites());
-        })
-        .catch((error) => {
-          dispatch(setFavouriteLoading(false));
-          dispatch(setFavouriteError(error));
-        });
-
-      return;
-    }
-
-    createFavourite(token, resource)
-      .then(() => {
-        // asks for the new updated list of favourites
-        dispatch(getUserFavourites());
-      })
-      .catch(({ errors }) => {
-        dispatch(setFavouriteLoading(false));
-        dispatch(setFavouriteError(errors));
-      });
-  });
 
 // Areas
 export const setUserAreas = createAction(SET_USER_AREAS);