+++ components/admin/data/widgets/form/steps/index.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 
-// constants
+// utils
 import {
   getRWAdapter,
 } from 'utils/widget-editor';
@@ -11,10 +11,5 @@ import Step1 from './component';
 export default connect((state) => ({
   user: state.user,
   query: state.routes.query,
-  RWAdapter: getRWAdapter({
-    locale: state.common.locale,
-    ...state.user.token && {
-      userToken: state.user.token.split(' ')[1],
-    },
-  }),
+  RWAdapter: getRWAdapter(state),
 }), null)(Step1);