+++ components/app/myrw/widgets/tabs/edit/index.js
@@ -11,10 +11,5 @@ import MyRWWidgetEditTab from './component';
 export default connect((state) => ({
   user: state.user,
   locale: state.common.locale,
-  RWAdapter: getRWAdapter({
-    locale: state.common.locale,
-    ...state.user.token && {
-      userToken: state.user.token.split(' ')[1],
-    },
-  }),
+  RWAdapter: getRWAdapter(state),
 }), null)(MyRWWidgetEditTab);