+++ layout/embed/dataset/index.js
@@ -12,7 +12,7 @@ export default connect(
   (state) => ({
     locale: state.common.locale,
     routes: state.routes,
-    RWAdapter: getRWAdapter({ locale: state.common.locale }),
+    RWAdapter: getRWAdapter(state),
   }),
   null,
 )(LayoutEmbedDataset);